fix(plugin): address pr comments
